RavenDB-23573 - Initialize `dynamicFieldsTerms` with correct comparer - CoraxJintConverter: in the case of dynamic vector search, we have to create a dynamic field but never persist it in the converter. - Moving methods related to VectorSearch from StaticIndexBase to its own separate file to improve readability - StaticIndexBase: correct detection of null values and strings - StaticIndexBase: added a common method for index field retrieval - StaticIndexBase: cleanup of VectorSearch methods
