+++ src/Raven.Server/Documents/Indexes/Static/StaticIndexBase.cs
@@ -17,6 +17,7 @@ using Raven.Client.Documents.Indexes.Vector;
 using Raven.Client.Documents.Linq.Indexing;
 using Raven.Server.Documents.Indexes.Persistence.Corax;
 using Raven.Server.Documents.Indexes.Persistence.Lucene.Documents;
+using Raven.Server.Documents.Indexes.Static.JavaScript;
 using Raven.Server.Documents.Indexes.Static.Spatial;
 using Raven.Server.Logging;
 using Raven.Server.Documents.Indexes.VectorSearch;
@@ -185,7 +186,7 @@ namespace Raven.Server.Documents.Indexes.Static
         }
     }
 
-    public abstract class AbstractStaticIndexBase
+    public abstract partial class AbstractStaticIndexBase
     {
         protected readonly Dictionary<string, CollectionName> _collectionsCache = new Dictionary<string, CollectionName>(StringComparer.OrdinalIgnoreCase);
 
@@ -342,425 +343,7 @@ namespace Raven.Server.Documents.Indexes.Static
             }
         }
 
-        /// <summary>
-        /// Dictionary training process occurs in IsOnBeforeExecuteIndexing. Since we're not training dictionaries with vectors, and considering computation
-        /// power required (e.g., generating embeddings from text), it is better to skip that part as there is no benefit in performing it.
-        /// </summary>
-        /// <param name="currentIndexingScope">Current indexing scope.</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static bool IsDictionaryTrainingPhase(CurrentIndexingScope currentIndexingScope)
-        {
-            return currentIndexingScope != null && currentIndexingScope.Index.IsOnBeforeExecuteIndexing;
-        }
-
-        internal static IndexField RetrieveVectorField(string fieldName, object value)
-        {
-            var currentIndexingScope = CurrentIndexingScope.Current;
-            var fieldExists = currentIndexingScope.Index.Definition.IndexFields.TryGetValue(fieldName, out var indexField);
-            if (fieldExists == false || indexField?.Vector is null)
-            {
-                // We're supporting two defaults:
-                // when Options are not set, we'll decide what is configuration in following manner:
-                // - value is textual or array of textual we're treating them as text input
-                // - otherwise, we will write as array of numerical values
-                var isText = IsText(value);
-                if (isText == false && value is IEnumerable enumerable)
-                {
-                    foreach (var item in enumerable)
-                    {
-                        if (item is null or DynamicNullObject)
-                            continue;
-                        
-                        isText = item is LazyStringValue or LazyCompressedStringValue or string or DynamicNullObject or JsString;
-                        break;
-                    }
-                }
-                
-                indexField = currentIndexingScope.GetOrCreateVectorField(fieldName, isText);
-            }
-
-            PortableExceptions.ThrowIf<InvalidDataException>(indexField?.Vector is null,
-                $"Field '{fieldName}' does not exist in this indexing scope. Cannot index as vector.");
-
-            indexField!.Vector!.ValidateDebug();
-
-            return indexField;
-            
-            bool IsText(object item)
-            {
-                return item is LazyStringValue or LazyCompressedStringValue or string or DynamicNullObject or JsString;
-            }
-        }
-        
-        public object CreateVector(string fieldName, object value)
-        {
-            if (value is null or DynamicNullObject)
-                return null;
-
-            var currentIndexingScope = CurrentIndexingScope.Current;
-            if (IsDictionaryTrainingPhase(currentIndexingScope))
-                return null;
-            
-
-            var indexField = RetrieveVectorField(fieldName, value);
-            var vector = indexField!.Vector!.SourceEmbeddingType switch
-            { 
-                VectorEmbeddingType.Text => VectorFromText(indexField, value),
-                _ => VectorFromEmbedding(indexField, value)
-            };
-            
-            if (indexField.Id != Corax.Constants.IndexWriter.DynamicField)
-                return vector;
-            
-            currentIndexingScope.DynamicFields ??= new Dictionary<string, IndexField>();
-            if (currentIndexingScope.DynamicFields.TryGetValue(fieldName, out var existing) == false)
-            {
-                currentIndexingScope.DynamicFields[fieldName] = indexField;
-                currentIndexingScope.IncrementDynamicFields();
-            }
-
-            return new CoraxDynamicItem() { Field = indexField, Value = vector };
-        }
-
-        /// <summary>
-        /// Create vector field object. This method is used by AutoIndexes and JavaScript indexes.
-        /// </summary>
-        /// <param name="indexField">IndexField from IndexDefinition</param>
-        /// <param name="value">Data source to create vector field.</param>
-        /// <returns></returns>
-        internal static object CreateVector(IndexField indexField, object value, bool isAutoIndex)
-        {
-            if (IsDictionaryTrainingPhase(CurrentIndexingScope.Current))
-                return null;
-            
-            return indexField!.Vector!.SourceEmbeddingType switch
-            {
-                VectorEmbeddingType.Text => VectorFromText(indexField, value),
-                _ => VectorFromEmbedding(indexField, value, isAutoIndex)
-            };
-        }
-
-        private static object VectorFromEmbedding(IndexField currentIndexingField, object value, bool isAutoIndex = false)
-        {
-            if (value is null)
-                return VectorValue.Null;
-            
-            var vectorOptions = currentIndexingField.Vector;
-            var allocator = CurrentIndexingScope.Current.IndexContext.Allocator;
-            if (value is LazyStringValue or LazyCompressedStringValue or string or DynamicNullObject)
-                return Base64ToVector(value);
-
-            if (value is BlittableJsonReaderArray || value is DynamicArray { Inner: BlittableJsonReaderArray })
-            {
-                var bjra = value as BlittableJsonReaderArray ?? (BlittableJsonReaderArray)((DynamicArray)value).Inner;
-                return HandleBlittableJsonReaderArray(bjra);
-            }
-
-            if (value is BlittableJsonReaderObject or DynamicBlittableJson)
-            {
-                var bjro = value as BlittableJsonReaderObject ?? ((DynamicBlittableJson)value).BlittableJson;
-                if (bjro.TryGetMember(Sparrow.Global.Constants.Naming.VectorPropertyName, out var vector) && vector is BlittableJsonReaderVector bjrv)
-                {
-                    return HandleBlittableJsonReaderVector(bjrv);
-                }
-
-                throw new ArgumentException($"Expected BlittableJsonReaderVector, but got {value.GetType().FullName}");
-            }
-
-            if (value is Stream stream)
-                return HandleStream(stream);
-            
-            if (value is JsArray js)
-                return HandleJsArray(js);
-
-            if (value is IEnumerable ie)
-                return HandleEnumerable(ie);
-            
-            throw new InvalidOperationException($"Unknown type. Value type: {value.GetType().FullName}");
-
-            object Base64ToVector(object base64)
-            {
-                var str = base64.ToString();
-                return GenerateEmbeddings.FromBase64Array(vectorOptions, allocator, str, isAutoIndex);
-            }
-
-            object HandleEnumerable(IEnumerable enumerable)
-            {
-                var enumerator = enumerable.GetEnumerator();
-                
-                using var _ = enumerator as IDisposable;
-                if (enumerator.MoveNext() == false)
-                    return null;
-
-                var isBase = IsBase64(enumerator.Current);
-                var isStream = enumerator.Current is Stream;
-                List<object> values = new();
-                
-                do
-                {
-                    ProcessItem(enumerator.Current);
-                }
-                while (enumerator.MoveNext());
-                
-                return values;
-
-                void ProcessItem(object item)
-                {
-                    if (isBase)
-                    {
-                        values.Add(Base64ToVector(item));
-                        return;
-                    }
-
-                    if (isStream)
-                    {
-                        values.Add(HandleStream((Stream)item));
-                        return;
-                    }
-                    
-                    switch (vectorOptions.SourceEmbeddingType)
-                    {
-                        case VectorEmbeddingType.Single:
-                        {
-                            var memScope = allocator.Allocate((((float[])item)!).Length * sizeof(float), out Memory<byte> mem);
-                            MemoryMarshal.Cast<float, byte>((float[])item).CopyTo(mem.Span);
-                            var vector = GenerateEmbeddings.FromArray(allocator, memScope, mem, vectorOptions, mem.Length);
-                            values.Add(vector);
-                            return;
-                        }
-                        case VectorEmbeddingType.Int8:
-                        {
-                            var memScope = allocator.Allocate((((sbyte[])item)!).Length, out Memory<byte> mem);
-                            MemoryMarshal.Cast<sbyte, byte>((sbyte[])item).CopyTo(mem.Span);
-                            var vector = GenerateEmbeddings.FromArray(allocator, memScope, mem, vectorOptions, mem.Length);
-                            values.Add(vector);
-                            return;
-                        }
-                        default:
-                        {
-                            var memScope = allocator.Allocate((((byte[])item)!).Length, out Memory<byte> mem);
-                            ((byte[])item).CopyTo(mem.Span);
-                            var vector = GenerateEmbeddings.FromArray(allocator, memScope, mem, vectorOptions, mem.Length);
-                            values.Add(vector);
-                            break;
-                        }
-                    }
-                }
-            }
-            
-            object HandleBlittableJsonReaderArray(BlittableJsonReaderArray data)
-            {
-                var dataLength = data.Length;
-                
-                if (TryGetFirstNonNullElement(data, out var firstNonNull) == false)
-                    return VectorValue.Null;
-                
-                //Array of base64s
-                if (IsBase64(firstNonNull))
-                {
-                    var values = new object[dataLength];
-                    for (var i = 0; i < dataLength; i++)
-                        values[i] = Base64ToVector(data[i].ToString());
-
-                    return values;
-                }
-                
-                //Array of arrays
-                if (firstNonNull is BlittableJsonReaderArray)
-                {
-                    var values = new object[dataLength];
-                    for (var i = 0; i < dataLength; i++)
-                    {
-                        if (data[i] == null)
-                            values[i] = VectorValue.Null;
-                        else
-                            values[i] = HandleBlittableJsonReaderArray((BlittableJsonReaderArray)data[i]);
-                    }
-
-                    return values;
-                }
-                
-                var bufferSize = dataLength * (vectorOptions.SourceEmbeddingType) switch
-                {
-                    VectorEmbeddingType.Single => sizeof(float),
-                    _ => sizeof(byte)
-                };
-
-                var memScope = allocator.Allocate(bufferSize, out Memory<byte> mem);
-                ref var floatRef = ref MemoryMarshal.GetReference(MemoryMarshal.Cast<byte, float>(mem.Span));
-                ref var sbyteRef = ref MemoryMarshal.GetReference(MemoryMarshal.Cast<byte, sbyte>(mem.Span));
-                ref var byteRef = ref MemoryMarshal.GetReference(mem.Span);
-                
-                for (int i = 0; i < dataLength; ++i)
-                {
-                    switch (vectorOptions.SourceEmbeddingType)
-                    {
-                        case VectorEmbeddingType.Single:
-                            Unsafe.Add(ref floatRef, i) = data.GetByIndex<float>(i);
-                            break;
-                        case VectorEmbeddingType.Int8:
-                            Unsafe.Add(ref sbyteRef, i) = data.GetByIndex<sbyte>(i);
-                            break;
-                        default:
-                            Unsafe.AddByteOffset(ref byteRef, i) = data.GetByIndex<byte>(i);
-                            break;
-                    }
-                }
-
-                return GenerateEmbeddings.FromArray(allocator, memScope, mem, vectorOptions, bufferSize);
-            }
-
-            object HandleBlittableJsonReaderVector(BlittableJsonReaderVector bjrv)
-            {
-                var type = bjrv[0].GetType();
-
-                if (type == typeof(float))
-                    return HandleBjrvInternal(bjrv.ReadArray<float>());
-                if (type == typeof(double))
-                    return HandleBjrvInternal(bjrv.ReadArray<double>());
-                if (type == typeof(byte))
-                    return HandleBjrvInternal(bjrv.ReadArray<byte>());
-                if (type == typeof(sbyte))
-                    return HandleBjrvInternal(bjrv.ReadArray<sbyte>());
-                
-                throw new NotSupportedException($"Embeddings of type {type.FullName} are not supported.");
-            }
-
-            object HandleBjrvInternal<T>(ReadOnlySpan<T> embedding) where T : unmanaged
-            {
-                var bufferSize = embedding.Length * (vectorOptions.SourceEmbeddingType) switch
-                {
-                    VectorEmbeddingType.Single => sizeof(float),
-                    _ => sizeof(byte)
-                };
-                
-                var memScope = allocator.Allocate(bufferSize, out Memory<byte> mem);
-                
-                MemoryMarshal.Cast<T, byte>(embedding).CopyTo(mem.Span);
-                
-                return GenerateEmbeddings.FromArray(allocator, memScope, mem, vectorOptions, bufferSize);
-            }
-
-            object HandleJsArray(JsArray jsArray)
-            {
-                var firstItem = jsArray[0];
-                if (firstItem.IsString())
-                {
-                    var values = new object[jsArray.Length];
-                    for (var i = 0; i < jsArray.Length; i++)
-                        values[i] = GenerateEmbeddings.FromBase64Array(vectorOptions, allocator, jsArray[i].AsString());
-
-                    return values;
-                }
-
-                if (firstItem is JsArray)
-                {
-                    var values = new object[jsArray.Length];
-                    for (var i = 0; i < jsArray.Length; i++)
-                        values[i] = HandleJsArray(jsArray[i] as JsArray);
-
-                    return values;
-                }
-
-                var len = (int)jsArray.Length;
-                var bufferSize = len * (vectorOptions.SourceEmbeddingType) switch
-                {
-                    VectorEmbeddingType.Single => sizeof(float),
-                    _ => sizeof(byte)
-                };
-
-                var memScope = allocator.Allocate(bufferSize, out Memory<byte> mem);
-                ref var floatRef = ref MemoryMarshal.GetReference(MemoryMarshal.Cast<byte, float>(mem.Span));
-                ref var sbyteRef = ref MemoryMarshal.GetReference(MemoryMarshal.Cast<byte, sbyte>(mem.Span));
-                ref var byteRef = ref MemoryMarshal.GetReference(mem.Span);
-                
-                for (int i = 0; i < len; ++i)
-                {
-                    var num = jsArray[i].AsNumber();
-                    switch (vectorOptions.SourceEmbeddingType)
-                    {
-                        case VectorEmbeddingType.Single:
-                            Unsafe.Add(ref floatRef, i) = (float)num;
-                            break;
-                        case VectorEmbeddingType.Int8:
-                            Unsafe.Add(ref sbyteRef, i) = Convert.ToSByte(num);
-                            break;
-                        default:
-                            Unsafe.AddByteOffset(ref byteRef, i) = Convert.ToByte(num);
-                            break;
-                    }
-                }
-                
-                return GenerateEmbeddings.FromArray(allocator, memScope, mem, vectorOptions, bufferSize);
-            }
-            
-            object HandleStream(Stream stream)
-            {
-                var len = (int)stream.Length;
-                var memScope = allocator.Allocate((int)stream.Length, out Memory<byte> mem);
-                stream.ReadExactly(mem.Span);
-                return GenerateEmbeddings.FromArray(allocator, memScope, mem, vectorOptions, len);
-            }
-            
-            bool IsBase64(object val) => val is LazyStringValue or LazyCompressedStringValue or string or DynamicNullObject or JsString;
-            
-            bool TryGetFirstNonNullElement(BlittableJsonReaderArray data, out object first)
-            {
-                first = data[0];
-                
-                var i = 0;
-                
-                while (first is null && i < data.Length)
-                    first = data[i++];
-                
-                if (first == null)
-                    return false;
-
-                return true;
-            }
-        }
-
-        private static object VectorFromText(IndexField indexField, object value)
-        {
-            if (value is null)
-                return null;
-
-            var allocator = CurrentIndexingScope.Current.IndexContext.Allocator;
-            object embedding = null;
-            if (value is LazyStringValue or LazyCompressedStringValue or string or DynamicNullObject)
-                embedding = CreateVectorValue(value);
-            else if (value is IEnumerable enumerable)
-            {
-                var vectorList = new List<VectorValue>();
-                foreach (var item in enumerable)
-                {
-                    vectorList.Add(CreateVectorValue(item));
-                }
-                
-                embedding = vectorList;
-            }
-
-            PortableExceptions.ThrowIf<InvalidDataException>(embedding is null, $"Tried to convert text into embeddings but got type {value?.GetType().FullName} which is not supported.");
-            return embedding;
-
-            VectorValue CreateVectorValue(object valueToProcess)
-            {
-                var str = valueToProcess switch
-                {
-                    LazyStringValue lsv => (string)lsv,
-                    LazyCompressedStringValue lcsv => lcsv,
-                    string s => s,
-                    DynamicNullObject => null,
-                    null => null,
-                    _ => throw new NotSupportedException("Only strings are supported, but got: " + valueToProcess.GetType().FullName)
-                };
-
-                if (str == null)
-                    return VectorValue.Null;
-
-                return GenerateEmbeddings.FromText(allocator, indexField.Vector, str);
-            }
-        }
+       
 
         public dynamic LoadDocument<TIgnored>(object keyOrEnumerable, string collectionName)
         {
